Rename project package and moved views to WEB-INF
