+++ WebContent/WEB-INF/cloudDrive/RegistrationView.jsp
