+++ src/cloudDrive/RegistrationController.java
@@ -1,4 +1,4 @@
-package filehosting;
+package cloudDrive;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
